mostly bugfixes and some refactoring

+++ include/sys/types.h
@@ -7,7 +7,7 @@
 
 typedef uint32_t blkcnt_t;
 typedef uint16_t blksize_t; // if anything wants a block size of over 64k it's Stupid and Dumb and i Don't Like It
-typedef uint8_t dev_t; // TODO: figure out how to properly handle this
+typedef size_t dev_t; // TODO: figure out how to properly handle this
 typedef uint16_t gid_t;
 typedef uint32_t ino_t;
 typedef uint16_t mode_t;
